Remove set_*_logger return values (#3909)
These return values are unused in the codebase.

Generally code uses a sub-logger of the logger used by these two
set_*_logger functions, mirroring the module hierarchy.

So it would be unusual for this logger return value to be used ...

... but this value *was* used in PR #3254 as part of ongoing refactoring
of logging helpers in other processes, until PR #3891, which
rationalised those other process logging to also be module hierarchy
based meaning the return value is not needed any more.

This is a tidyup to allow a future commit to return a different value as
part of work to shut down parsl logging at DFK shutdown.

# Changed Behaviour

Users making use of this return value in their own code will have to
acquire the relevant `logger` through some other means - for example,
using the usual `logging.getLogger` idiom.

## Type of change

- Code maintenance/cleanup