Increase log level for binding between WQ and executor ID namespaces (#3912)
This is to help understand the connection between Work Queue and
executor level tasks when looking through logs.

Awkardly this binding is only logged at the end of a task execution, but
this is intended to be a minimal PR.

See for example #2313 for similar binding work in the DataFlowKernel.

# Changed Behaviour

one extra log line per task when running at INFO log level and using the
Work Queue executor.

## Type of change

- Update to human readable text: Documentation/error messages/comments