Move DB manager exit check happen to main thread
Prior to this PR, self.close() was called inside the message migration
thread. But self.close() does database stuff, which should all happen
in one thread: the main thread.

The historical reason is that self.close was originally triggered by
a special stop message in what was previously many threads all running
this migrate code. That stop message was removed in PR #3807 and replaced
by a multiprocessing Event.

This addresses some exceptions at shutdown when both the main thread and
the message migration close() attempt to write to the database at the
same time.

This PR does not address the possibility of outstanding messages which
have not reached the database manager at the time of shutdown.

The message migration thread now does only message routing.