Make executor.connected_workers a regular method (#3917)
See PR #3915 for related justification on the same change for
executor.outstanding()

In the connected_workers case, for High Throughput Executor, an RPC to
the interchange is made when this attribute is accessed, which is
execptionally high failure risk and has personally caused me a lot of
trouble in understanding previously.

# Changed Behaviour

User code that is accessing connected workers will have to switch to
method syntax, from attribute syntax.

## Type of change

- Code maintenance/cleanup