Replace use of queue.qsize() in monitoring database manager
Prior to this PR, each use of .qsize() was of the form

q.qsize() != 0

which is equivalent to (not q.empty())

Queue.empty() is better supported across platforms than .qsize()

This PR makes that substitution, meaning the broken (#3856) MacSafeQueue
subclass will not be needed in this situation any more and the database
manager will be able to work with a regular multiprocessing Queue even
on Mac.

This PR does not actually change the Queue type - it only reduces the
functionality required by whichever queue is supplied.

This is part of a series of PRs to remove that platform-specific
functionality entirely - see PR #3932.