Add regression test for htex command client concurrency fix
The concurrency fix is in PR #1321, around 6 years ago.

With this long hindsight, the corresponding issue #1146 is a
concurrency problem in use of the ZMQ REQ/REP socket used by the
HTEX command client; and PR #1321 is a fix for that concurrency.

After that realization, it was straightforward to build this concurrency
test which shows failure within a couple of seconds when #1321 is removed,
on both my laptop and in GitHub CI.

I'm adding this test now because I would like to simplify the command
client code that this tests in upcoming PRs.