Reduce SizedQueue to a regular Queue in monitoring radios (#3936)
The radios do not make use of the .qsize() functionality of queues, and
so th (broken, see #3856) platform-specific sized queue functionality is
not necessary.

This is part of a series of PRs to remove that platform-specific
functionality entirely - see PR #3932.

# Changed Behaviour

This code is part of broken behaviour on Macs. The behaviour will be
slightly better but still broken after this PR on that platform.

Otherwise, the queue should not behave any differently.

## Type of change

- Code maintenance/cleanup