Remove manual checkpoint() directory return value (#3940)
Prior to this PR, a manual checkpoint() call returned the directory that
the checkpoint was written to.

This isn't really needed, so this PR removes it as part of removing
cruft around the checkpoint API, as part of refactoring towards
pluggable checkpointing.

The checkpoint directory isn't specific to manually-invoked
checkpointing so this is not the right way to be informing the user of
the directory.

The only use is in a test case, which this PR rewrites.

# Changed Behaviour

Anyone who expects to get the directory from a manual checkpoint call
will have to do so in some other way. See the changed test in this PR
for an example.

## Type of change

- Code maintenance/cleanup