Replace SizedQueue use in monitoring with regular Queue (#3944)
Since PR #3937, no .qsize() has been used in monitoring.

This leaves MacSafeQueue, the Mac implementation of SizedQueue, unused.
It will be removed in a subsequent PR. See PR #3932 for over-arching
work

# Changed Behaviour

Monitoring should now work on Mac.

# Fixes

Fixes #3856 - hopefully. @d33bs has tested an approximately-equal change
in #3932 on Mac, but that is not an auto-tested platform.

## Type of change

- Bug fix