Refactor packing and unpacking with resource spec
The fundamental change is:

```diff
- f, a, k, resource_spec = unpack_res_spec_apply_message(...)
+ f, a, k = unpack_apply_message(...)
```

in `execute_task()`.  This re-generalizes the `execute_task()` implementation
for all executors. From there, everything else is "cleanup" to not pack the
resource specification within the function buffer, or in the HTEX case, to use
the `resource_spec` passed along-side.

This removes the need entirely for `res_spec` variants of
`pack_res_spec_apply_message` and `unpack_res_spec_apply_message`.  The
requisite functionality (`for`-loop) is now in the HTEX's worker
(`process_worker_pool.py`).