Remove htex command client retry logic (#3946)
This retry logic was introduced as a bug fix in #1321 alongside a new
lock around the command client channel.

On re-inspection of the issue (#1146), to me it looks like the bug is
un-controlled multithreaded use of the command client ZMQ channel; and
looks like #1321 solves this by introducing that new lock, not by
introducing the retry logic, despite the description and bulk of body of
PR #1321 refering to the retry mechanism.

In testing, I removed the lock and tried out the retry logic unprotected
- to see what happens when the retry logic does fire - and this resulted
in various hangs, segfaults and other exceptions. The new test in PR
#3935 is basically the test I used for this.

So, #1321 "should have" introduced the lock and not the retry logic.

This PR removes that retry logic as broken and unused, hopefully leaving
the behaviour essentially unchanged but with easier to read code.

# Changed Behaviour

This removes code around Parsl's traditionally race-prone ZMQ usage - so
there is a chance the peturbation reveals some new occasional race.

## Type of change

- Code maintenance/cleanup