Improve priorities test (#3955)
# Description

While working through a recent refactor, realized that I wasn't sure if
the list was just reversed or if there was another bug. Make it clearer
for the next interaction that it starts life as an intentionally
unsorted list, and the engine is supposed to deal with it correctly.

# Changed Behaviour

Only a test; no user-facing changes

## Type of change

- Code maintenance/cleanup