Don't hide original exception (#3963)
# Description

The caught exception provided negligible value. Just remove it -- which
is, pragmatically speaking, the same behavior (we were already
throwing!), but now let the raw exception bubble rather hiding the
origin.

# Changed Behaviour

Pragmatically no difference, but in the case of inter-component serde
mismatch or bitrot, now emit a full traceback rather than only a partial
one with an unhelpful summarization.

## Type of change

- Update to human readable text: Documentation/error messages/comments
- Code maintenance/cleanup