Add conference proceedings paper references to `CITATION.cff` and `README.rst` (#3943)
# Description

This PR adds the conference proceedings paper on Parsl to the
`CITATION.cff` file and also adds a badge to the project to make the
paper more visible to others. I recognize there is also a `CITATION.md`
with this information but thought it could be helpful to raise further
awareness through more direct integration.

# Changed Behaviour

People can reference the `CIATION.cff` or the `README.rst` to find the
paper.

# Fixes

Closes #3857 

## Type of change

Choose which options apply, and delete the ones which do not apply.

- Update to human readable text: Documentation/error messages/comments