Fix bad global default provider in several executors
This affects the High Throughput Executor, the MPI Executor and
Task Vine.

Mutable objects should not be used as default values, and in this case,
a mutable LocalProvider is used as the default provider.

This PR changes that to use a `None` default value and makes a new
LocalProvider whenever a new instance of one of the relevant Executors
is made.

Fixes issue #3874.

See issue #1662 for broader context around mutable defaults.

This PR adds a regression test based on the reproducer supplied by
@d33bs in issue #3874.