Make slurm test container use virtualenv
As part of removing support for Python 3.9, the slurm
test container was updated (not in this repo) to use a more
recent OS/Python version. That more recent version is
more aggressive about protecting against root-level global
installation of software using `pip` rather than using the
OS-native package manager.

This PR fixes that by using a virtualenv, which is consistent
with the behaviour in the main CI environment.