TODO: factor a complete_exception method, like complete_task. and make all this stuff happen inside those?
TODO: should these run in the app_fu update lock? there's a bunch of other completion stuff that is not
locked, so probably no? in general this lock isn't protecting against two completion actions happening
at once.

split update_memo into two notify-exception and notify-result forms, that run before the future is set and which
means checkpointing can run in that place rather than racing with the user workflow.

this is needed to address an existing race condition - that task_exit tasks aren't checkpointed by the time
the workflow observes they have finished. it's debatable whether that is actually the intended functionality
or not, but i think it should be. otherwise what's the behaviour? a little bit later?

this manifests as awkwardness in implementation in later checkpoint methods too.

and moving more checkpoint code into update_memo makes the checkpoint/memo API unification cleaner.

See PR #NNNN which talks about the app future being used internally when it should only be user-facing.