Move checkpointing earlier in task completion
Prior to this PR, checkpointing happened in the inherently race-prone
handle_app_update where the user has already been informed that the
task has completed. This means the user can observe a situation where
checkpointing has not happened even though the task appears to have
completed.

Because the user-facing app future is now not set by the time of
checkpointing, this PR changes the checkpointing code to pass results
and exceptions explicitly - following on from PR #4017.