Remove special-case checkpoint case from garbage collector test
Prior to PR #2514, tasks were removed at completion when checkpointed,
rather than when completed, in the case that checkpointing was enabled.

This was race-y behaviour and was removed in PR #2414.

This test also looks like it is subject to a related race condition
around task completion, described in issue #1279, for which there is a
sleep statement - this PR does not modify that but does add a reference
to the issue.