Remove support for Python 3.9 (#3966)
# Description

Python 3.9 becomes end-of-life in the next month - see
https://devguide.python.org/versions/

This PR executes Parsl's informal policy to drop Python version support
around the time that a Python version becomes end-of-life.

At the same time and on a similar topic, PR #3924 investigates adding
support for Python 3.14 which has an initial release scheduled soon.

# Changed Behaviour

If you are installing on Python 3.9, you will probably now get package
install errors that your Python is too old.
Post this PR, you will start finding the codebase begin to use features
which are not supported on Python 3.9 without any further notification.

## Type of change

- Code maintenance/cleanup

---------

Co-authored-by: Kevin Hunter Kesling <kevin@globus.org>