Add configurable iteration modes to parsl-perf (#3985)
These are described in the documentation added in this PR.

The default mode is the previous behaviour (estimate mode).

This PR adds a new mode, exponential, which is useful for looking at
scaling behaviour over batch sizes and for more consistent batch sizes
across runs.

# Changed Behaviour

new options, old behaviour should be preserved

## Type of change

- New feature