Remove requirement that executors implement their own .start() (#3987)
This is most immediately driven by my work to upgrade to mypy 1.18.2
from mypy 1.5.1. In that change, mypy now raises an error if a class
calls the abstract-decorated superclass method:

```
parsl/executors/globus_compute.py:81: error: Call to abstract method
"start" of "ParslExecutor" with trivial body via super() is unsafe  [safe-super]
```

This happens elsewhere in the codebase too, but mypy doesn't notice
those occurences.

As demonstrated by the GlobusComputeExecutor, it isn't actually
mandatory for executor-specific start behaviour to exist.

So, this PR removes the @abstractmethod decorator from the
executor.start() method. It is no longer mandatory to implement; the
existing no-op default implementation remains but no longer upsets more
recent mypy. The .start method remains as an optional hook which most
executors still choose to use.

This PR also removes the stub in GlobusComputeExecutor which was there
only to satisfy the now-removed @abstractmethod requirement.

# Changed Behaviour

Shouldn't change user experience. Developers of new executors will no
longer be forced to implement start and will have to discover it (if
they want it) by reading.

## Type of change

- Code maintenance/cleanup