Don't strategize executors in bad state
This is consistent with JobStatusPoller.close, which does not scale in
executors in bad state.

See issue #3992 for more context. This fixes a race condition that I can
create by adding fuzz delays (as described in #3992) and that appeared
(without adding fuzz timing) in PR #3991 probably due to reduced time
taken per strategy iteration in that PR.