Address race condition in htex priority test (#3998)
This race condition is in the test, not in the main Parsl codebase, and
is that the test was not waiting for tasks to all arrive in the
interchange before launching a worker.

This was happening fast enough most of the time, but under some extreme
race condition testing, this was breaking the test - I'm working towards
that testing being automated in CI, so this is not a theoretical
concern.

If the tasks are not all in the interchange when a worker connects, they
won't necessarily be run in priority order: the highest priority task of
the subset of tasks that has already arrived will be run first, even if
it is not the highest priority task of the entire task set.

# Changed Behaviour

this is a test behaviour change, should not be user-exposed

## Type of change

- Bug fix