Pull more completion code into _complete_task and rename
This PR adds _send_task_log_info as a mandatory part of task completion,
which is already manually enforced. In one place, around line 500, one
_send_task_log_info is split into two: one which then disappears into
complete_task_result, and the other which remains in place for the
non-factored error handling path.

This PR renames _complete_task to _complete_task_result to reflect that
it is for the result path not the exception path (taking naming from
the Future concepts of the same name), and to open up the way for
factorisation of the exception path into _complete_task_exception.