Replace an always-true `if` with an assert
The "else" case of the erstwhile-if statement was not defined properly:
What are the circumstances and behaviour if an exec_fu is not set by
this point?

All don't-launch-yet behaviour is covered by `return` statements earlier
on and the second part of the method should always proceed with making
an exec_fu of some kind.