Don't strategize executors in bad state (#3994)
This is consistent with JobStatusPoller.close, which does not scale in
executors in bad state.

See issue #3992 for more context. This fixes a race condition that I can
create by adding fuzz delays (as described in #3992) and that appeared
(without adding fuzz timing) in PR #3991 probably due to reduced time
taken per strategy iteration in that PR.

The new use of `bad_state_is_set` is prone to race conditions - that's
part of the model of how bad_state_is_set works everywhere in the
codebase, and is briefly discussed in PR #3995.

# Changed Behaviour

Scaling will not happen for bad-state executors. I think this is always
the right thing to do.

## Type of change

- Bug fix