Remove connected workers call in strategy (#3991)
This isn't used for strategizing. It is an RPC into the htex interchange
that is at elevated risk of thread-related ZMQ misuse failures. For
other executors, it is ignored.

It is an abstraction leakage in general: nothing in the job scaling code
needs to know about connected workers, and executors in general don't
have a per-task-slot concept of connected workers. Even the MPI variant
of HTEX doesn't have that.

This removes connected workers use entirely, but leaves the call in
place for interested users (and tests).

The interchange log already contains details of workers, and that is
where users should now go for worker information.

This makes debugging a little different: this information was useful to
see periodically in Parsl logs to see if workers are connected, and now
that information lives elsewhere: in interchange.log.

However the strategy polling thread is not a centralised status/logging
aggregator system!

# Changed Behaviour

Less logging information in main parsl.log about connected workers -
none at all, in fact. Balancing that, hopefully less ZMQ-misuse-related
misbehaviour.

## Type of change

- Code maintenance/cleanup