Remove bash app exception return code report that is constant None (#4008)
There is no code path where this value gets set and then ends up in the
exception block, and this code misleading leads a reader to think the
unix return code will be reported here.

The unix return code is actually reported for failing apps a few lines
later with a BashExitFailure exception.

# Changed Behaviour

The text "with returncode: None" no longer appears on this exception.

## Type of change

- Update to human readable text: Documentation/error messages/comments