Refactor failure states collection (#4009)
# Description

Reviewing PR #4006 brought up the fact that the `FINAL_FAILURE_STATES`
is a strict subset of `FINAL_STATES`. The thrust of this commit is to
create `FINAL_STATES` as a superset of `FINAL_FAILURE_STATES`.

While here, favor a Python `set()` (`{...}`) for the existence criteria
over `list()` (`[...]`). I do not intend this as any sort of performance
improvement but as a matter of correctness: these objects are about
existence, so (unique) sets make more natural sense.

# Changed Behaviour

none

## Type of change

- Code maintenance/cleanup