Implement exception version of complete_task_result factorisation (#4006)
The new complete_task_exception and the older complete_task_result share
a lot of code -- but upcoming PRs in my current checkpoint work will
make them behave a bit differently later, so I don't want to merge them
now and unmerge them later.

The order of task completion operations is not preserved by this PR:
before this PR, different instances of the exception completion boiler
plate ran the same operations in different orders.

I don't expect that to be a problem, but if you've bisected to this PR
with a strange ordering/race condition, there's a clue.

This PR removes a comment from one instance of the boilerplate that asks
a question fixed by a previous PR #4004 that also accidentally
introduced the question while tidying it.

# Changed Behaviour

none

## Type of change

- Code maintenance/cleanup