move update status next to task info send
this reordering mostly should be ok

but i am getting some kind of error in another patch
so do it explicitly here.

around not parsing AUTO_LOGNAME as -1
(which is the value of AUTO_LOGNAME)

after this PR:

every update task state and send_task_info pair should
be right next to each other, update first then send_task_info next

in preparation for refactoring

and there should be no calls to one without the other.

this PR has identified one such situation where this isn't the
case, which causes an exception log message - that drives a previous
patch to remove the unsched state entirely.