make tasks have their own loggers in TaskRecord
right now these should be adapters onto the DFK logger

as discussed in immediately preceeding patch

but they're perhaps also the API point for giving a per-task
observability object in memory that other things can happen to?

this makes the code a bit more verbose by line count
(eg 39 removed, 69 added)
but the point is to get objects in there for more general
observability, not be to concise in code.