Add more unit tests in the `get_app` function, document chaining (#4014)
# Description

I wanted to make sure that Parsl futures created with ParslPoolExecutor
work with Parsl function chaining.
As expected, they do. These changes make the compatibility between PPE
and other Parsl routes clear in the documentation too

# Changed Behaviour

No changes in functionality.

I did make a method public so the unit tests were clearer and open that
route as an acceptable option in a workflow I'm tinkering with.

# Fixes

N/A

## Type of change

- Update to human readable text: Documentation/error messages/comments