Work towards new mypy, 1.18.2 (#4007)
This PR has the usual upgrade-mypy dance of removing and adding various
ignores as mypy changes what it can reason about.

This PR removes one use of AnyStr which was incorrect: it is a type
variable for aligning multiple uses of str|bytes together, but it was
not properly used in that was in the removed case.

The immediate motivation for this PR is to bring in newer mypy support
for match statements, as part of work on parsl-perf which uses match
statements now.

# Changed Behaviour

none

## Type of change

- Code maintenance/cleanup