the current memoizer.checkpoint call is a bit overloaded
split it

although there is an obvious future-based way of refactoring,
the implementations will diverge later: queue based checkpointing
will use futures, but checkpoint_one will use result/exception
style - and probably become checkpoint_one_result, checkpoint_one_exception?

further, there's new race behaviour if doing it that way:
checkpoint_queue then relies on the task future being done.
so the ordering about adding a future to the queue vs
setting the result is horrible again... adding it to the
queue would have to accept that the result might not be
done yet and it should remain queued for later.
but then calling checkpoint() after a task won't guarantee
that it gets checkpointed. i think both ordering are a bit
horrible.

ugh

completion semantics

or make the queue not use futures at all and instead record the exception/result as a two element xor-tuple emulating a sum type - that's more in line with the actual principles of how checkpointing should work.

is this patch the right place to do it? yes, because the next PR does that sort of stuff

exception checkpointing doesn't appear until much later.