Remove GCE test (#4025)
Parsl's `GlobusComputeExecutor` code is apparently unmaintained. Rather
than muddy all other PRs with the failing test -- even if it's not
required for merging -- be the better netizen and just remove the test.
At such time as this plugin is updated, we can revisit the test.

# Changed Behaviour

Removes the now-consistently failing GCE tests, so no more failing test
for all other PRs.

## Type of change

- Code maintenance/cleanup