Rearrange task unscheduled state (#4023)
Removes the not-very-used unsched task status as the task is set to
pending later in the submission process: the only time anyone would see
this status is inspecting self.tasks.items() (or similar) for a small
window of time before that task is set to pending.

Adds the task record to self.tasks much later in the task construction
process, so that more of the record has been constructed by the time the
record is exposed that way.

This allows future PRs to make more assumptions about the state of the
task record when the task state is set/changed.

# Changed Behaviour

none

## Type of change

- Code maintenance/cleanup