hacking around with what configuration might look like.
use cases:

i) file based text logging - the current situation

- with configurable debug and topic levels - currently set by ad-hoc debug options


ii) json logging - alternative or in addition to the current situation

iii) logging to non-Parsl-included handler - for example an external kafka or chronolog

emulate the monitoring radio config style:

i) log config is given a chance to populate itself from the local environment (for working directory, credentials, ...)

ii) log config is now pickle-able

iii) log config needs to be conveyed and executed at the start of every python process. which is ad-hoc because they are all over the place. logging is process-global so perhaps I shouldn't be trying to do this at parsl.load time, and instead have the user perform global logging config for the submit process? because there isn't a notion of per-dfk logging through the system at the moment (and there isn't on the horizon either)

that's a bit of a config shift from parsl.load() style configuration. and maybe actually parsl.load, being a global function, is allowed to perform non-DFK configuration too!

it it is thusly stored in the global environment (alongside the global (distributed among objects) Python-side log configuration) then anyone can grab it for use.

I would also like this approach to be somewhat modular with eg Academy and the other bits of the ecosystem I'm working with.


Note that some processes launched right now don't have logs at all (since I screwed that up before) - I think TV/WQ spawn processes - but anything that turned from a fork to a spawn process without consideration for logging

How to pass the pickle around for the process worker pool? Precedent: the certificates stuff is sent around as a secret in a file in a shared directory. The log configuration object may also contain secrets, and so should be treated as a secret too. Perhaps lets dump it in the certificates directory for now to emphasise that secretness?

But this is going to force the certificate directory to be more mandatory: which is not the case with kubernetes right now (and why people run kubernetes with no encryption...) -- OR i make a helper function that configures the traditional log file if nothing is passed?

initialize_logging gets these extended meaning:

True: initalise parsl.log and all subcomponent logs
False: do not initialize parsl.log. do initialize all subcomponent logs -- that's the behaviour now
callback: use the callback