Move checkpointing earlier in task completion (#4026)
Prior to this PR, checkpointing happened in the inherently race-prone
handle_app_update where the user has already been informed that the task
has completed. This means the user can observe a situation where
checkpointing has not happened even though the task appears to have
completed.

Because the user-facing app future is now not set by the time of
checkpointing, this PR changes the checkpointing code to pass results
and exceptions explicitly as a CheckpointCommand object. This aligns
with the on-disk checkpoint format and with the explicit separation of
results and exceptions in earlier work - eg PR #4017

This new CheckpointCommand object contains support for representing
exceptions because coming up in ~6 PRs time is support for choosing to
checkpoint exceptions as well as results, as a user-configurable policy.

See #1279 for summary issue on task completion, broader than
checkpointing.

# Changed Behaviour

checkpoint now happens at a different time. previously in task_exit
mode, users could observe tasks that were completed but not
checkpointed. that should now not be observable.

## Type of change

Choose which options apply, and delete the ones which do not apply.

- Bug fix
- New feature
- Update to human readable text: Documentation/error messages/comments
- Code maintenance/cleanup