apt update before installing slurm ci dependencies (#4039)
Prior to this PR, the SLURM CI build installed some dependencies via
`apt`, but the base container could become out of date wrt the live
repos. This PR does an `apt update` before that install. That's the
pattern I use in my own containers to address this drift.

The SLURM CI test is broken right now because of this drift - this PR
fixes it. (compare the red vs green test results on this PR vs on other
recent PRs).

cc @tylern4 for opinion as original contributor

# Changed Behaviour

none

## Type of change

- Code maintenance/cleanup