Define interface between DFK and memoizer (#4033)
This PR makes an interface class that defines the calls the DFK can make
to the memoizer, distinct from the implementation of the implementation
of the memoizer which is renamed to BasicMemoizer.

The DFK still knows about how to construct a BasicMemoizer, but after
that, self.memoizer is type-checked as the new Memoizer interface class.

A future PR will separate out that configuration so that the user
constructs a BasicMemoizer or some other plugin, using Parsl plugin
style. At the same time, this new interface will have its
checkpoint_queue method removed as being specific to the BasicMemoizer,
rather than memoizer-generic.

# Changed Behaviour

none

## Type of change

- Update to human readable text: Documentation/error messages/comments
- Code maintenance/cleanup