Switch two visualization views to safer SQL parameter styleA
Prior to this PR, these two SQL statements were formed by direct string
substitution, which allowed arbitrary text injection into the SQL string.

I think in the default configuration this isn't a security vulnerability,
because whatever is injected at these points is limited in what it can
do: the database is already public because thats what parsl-visualize
does, and sqlite will not allow other commands to be executed alongside
the query.

This was reported by @viralvaghela in
https://github.com/Parsl/parsl/security/advisories/GHSA-f2mf-q878-gh58
(unpublished at time of commit)